Update de clases

+++ Estrella.java
@@ -1,7 +1,7 @@
 public class Estrella extends ObjetoCeleste {
     private String tipoEspectral;
 
-    public Estrella(String nombre, String fecha, float magnitudAparente, String tipo, String telescopio, int[] coordenadas, String tipoEspectral) {
+    public Estrella(String nombre, String fecha, float magnitudAparente, String tipo, String telescopio, float[] coordenadas, String tipoEspectral) {
         super(nombre, fecha, magnitudAparente, tipo, telescopio, coordenadas);
         this.tipoEspectral = tipoEspectral;
     }
